Added "snap" to UI Helpers
Snap can be used for various purposes such as having snappable floating buttons on top of the UI without negatively effecting the UX or controlling the organisation of in-line elements while allowing the user to customise them to some extent as well. It is a very open-ended package and I believe it can be helpful to many Flutter coders.