Added "cupertino-tabbar" to Cupertino Design
I think the Flutter team has been a little late so far for adapting more to the new iOS 13  style design elements. So I created two new packages for helping:D This is one of them and it provides a widget for the new tab bars used by iOS 13. As you can see in the relevant repo, its very simple and very customisable.