Add customizable_date_picker
You've read How to contribute right ? Yes!

customizable_date_picker is a Dart Package that helps you build a Date range picker that's suits your needs in term of UI and functionalities.

Why did I create customizable_date_picker?

I was working in a project of House Renting system and i found myself stuggeling on creating a date range picker that suits the theme i am using and that gives the possibilty to show prices in each day. And after i started working on the DatePicker i realized how time consuming it is to make it from scratch, so i decided to make it as generic as possible so it can be used in other project.

It didn't take me more than 15min to build each of the examples i mentioned in the Github repo using the Package.